+++ training_pipeline_edna_gpu.py
@@ -1,37 +1,30 @@
+'''
+use : python training_pipeline_edna_gpu.py <parfile>.json
+'''
 
 import os
 import sys
-import json
+#import json
 import logging
 import argparse
 
 from dotenv import load_dotenv
+import infer.modules.train.training_pipeline as pipeline 
+
 now_dir = os.getcwd()
 sys.path.append(now_dir)
 load_dotenv()
 
-import infer.modules.train.training_pipeline as pipeline 
-
-######################################################################
 parser = argparse.ArgumentParser(
                     prog='Training Pipeline',
                     description='Runs training pipeline for RVC',
                     epilog='Ask me for help')
 parser.add_argument('parfile') 
 
-def read_param_dict(parser):
-    args = parser.parse_args()
-    json_file = open(args.parfile)
-    param_dict = json.load(json_file)
-    json_file.close()
-
-    return param_dict
-
-######################################################################
 
 logger = logging.getLogger(__name__)
 
-param_dict = read_param_dict(parser)
+param_dict = pipeline.read_param_dict(parser)
 config_vars = pipeline.read_config_vars()
 
 pipeline.preprocess_dataset(param_dict,
